requirements.txt
openai
pytelegrambotapi
python-dotenv
flask
